Added confirmation popup for track/layer deletion
Closes #45

+++ www/app/layers/layer.controller.js
@@ -5,10 +5,10 @@
         .module('TractNotes')
         .controller('LayerController', LayerController);
 
-    LayerController.$inject = ['$rootScope', 'layerControlService', 'layerViewService', 'ctecoDataService', 'wmsUrlService'];
+    LayerController.$inject = ['$rootScope', 'layerControlService', 'layerViewService', 'ctecoDataService', 'wmsUrlService', '$ionicPopup'];
 
     /* @ngInject */
-    function LayerController($rootScope, layerControlService, layerViewService, ctecoDataService, wmsUrlService) {
+    function LayerController($rootScope, layerControlService, layerViewService, ctecoDataService, wmsUrlService, $ionicPopup) {
         var vm = this;
         vm.title = 'TrackController';
 
@@ -36,17 +36,36 @@
 
         function sendLayerDelete(selectedLayer) {
             console.log('Layer type: ' + selectedLayer.layerType);
-            if (selectedLayer.layerType == 'cteco' || selectedLayer.layerType == 'ortho') 
-            {
-                ctecoDataService.deleteLayer(selectedLayer);
-            }
-            else if (selectedLayer.layerType == 'wmsTile' || selectedLayer.layerType == 'Dynamic Map Layer' ||
-                selectedLayer.layerType == 'ESRI Image Map Layer' || selectedLayer.layerType == 'ESRI Feature Layer' ||
-                selectedLayer.layerType == 'Tile Layer')
-            {
-                wmsUrlService.deleteLayer(selectedLayer);
-            }
-            selectedLayer.checked = false;
+            // confirmation popup for layer deletion
+            var confirmPopup = $ionicPopup.show({
+                title: 'Confirm Layer Deletion',
+                template: 'Are you sure you want to delete this layer (' + selectedLayer.name + ')?',
+                buttons: [
+                    {
+                        text: 'Delete',
+                        type: 'button-positive',
+                        onTap: function(e) {
+                            console.log('Layer deletion (' + selectedLayer.name + ') confirmed');
+                            if (selectedLayer.layerType == 'cteco' || selectedLayer.layerType == 'ortho') 
+                            {
+                                ctecoDataService.deleteLayer(selectedLayer);
+                            }
+                            else if (selectedLayer.layerType == 'wmsTile' || selectedLayer.layerType == 'Dynamic Map Layer' ||
+                                selectedLayer.layerType == 'ESRI Image Map Layer' || selectedLayer.layerType == 'ESRI Feature Layer' ||
+                                selectedLayer.layerType == 'Tile Layer')
+                            {
+                                wmsUrlService.deleteLayer(selectedLayer);
+                            }
+                            selectedLayer.checked = false;                            
+                        }
+                    },
+                    {
+                        text: 'Cancel',
+                        onTap: function(e) {
+                            console.log('Layer deletion (' + selectedLayer.name + ') canceled');
+                        }}
+                ]
+            });
         }
     }
 })();
